+++ integrations/llama_cpp/src/haystack_integrations/components/generators/llama_cpp/chat/chat_generator.py
@@ -86,6 +86,10 @@ def _convert_message_to_llamacpp_format(message: ChatMessage) -> ChatCompletionR
         if tool_call_results[0].origin.id is None:
             msg = "`ToolCall` must have a non-null `id` attribute to be used with llama.cpp."
             raise ValueError(msg)
+        # handle list of multimodal tool call results
+        if not isinstance(tool_call_results[0].result, str):
+            msg = "ToolCallResult.result must be a string to be used with llama.cpp."
+            raise ValueError(msg)
         return {
             "role": "function",
             "content": tool_call_results[0].result,